# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/portfolio/locales": {
      "get": {
        "operationId": "PortfolioController_getLocales",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      },
      "post": {
        "operationId": "PortfolioController_createLocale",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocaleDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 409
                    },
                    "error": {
                      "type": "string",
                      "example": "Conflict"
                    },
                    "message": {
                      "type": "string",
                      "example": "Locale already exists"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/portfolio/profile": {
      "get": {
        "operationId": "PortfolioController_getProfile",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ko",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unsupported locale"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Profile not found"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      },
      "put": {
        "operationId": "PortfolioController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/portfolio/profile/all": {
      "get": {
        "operationId": "PortfolioController_getProfileWithTranslations",
        "parameters": [],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 404
                    },
                    "error": {
                      "type": "string",
                      "example": "Not Found"
                    },
                    "message": {
                      "type": "string",
                      "example": "Profile not found"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/portfolio/experiences": {
      "get": {
        "operationId": "PortfolioController_getExperiences",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ko",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unsupported locale"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      },
      "post": {
        "operationId": "PortfolioController_createExperience",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperienceDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/portfolio/projects": {
      "get": {
        "operationId": "PortfolioController_getProjects",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ko",
              "type": "string"
            }
          },
          {
            "name": "featured",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unsupported locale"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      },
      "post": {
        "operationId": "PortfolioController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/portfolio/skills": {
      "get": {
        "operationId": "PortfolioController_getSkills",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      },
      "post": {
        "operationId": "PortfolioController_createSkill",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      }
    },
...
    "/api/portfolio/profile/image": {
      "post": {
        "operationId": "PortfolioController_uploadProfileImage",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "No file provided or invalid file type"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 401
                    },
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid credentials"
                    },
                    "path": {
                      "type": "string",
                      "example": "/api/..."
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2026-01-01T00:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "portfolio"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hyunwoo API",
    "description": "chahyunwoo.dev blog & portfolio API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "CreateLocaleDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "zh"
          },
          "label": {
            "type": "string",
            "example": "中文"
          }
        },
        "required": [
          "code",
          "label"
        ]
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "iconUrl": {
            "type": "string"
          },
          "socialLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialLinkDto"
            }
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileTranslationDto"
            }
          }
        }
      },
...
      "CreateExperienceDto": {
        "type": "object",
        "properties": {
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "isCurrent": {
            "type": "boolean",
            "default": false
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExperienceTranslationDto"
            }
          }
        },
        "required": [
          "startDate",
          "translations"
        ]
      },
...
    }
  }
}